+++ strawberry/subscriptions/__init__.py
@@ -1,2 +1,2 @@
-GRAPHQL_TRANSPORT_PROTOCOL = "graphql-transport-ws"
+GRAPHQL_TRANSPORT_WS_PROTOCOL = "graphql-transport-ws"
 GRAPHQL_WS_PROTOCOL = "graphql-ws"
