+++ tests/aiohttp/test_view.py
@@ -1,24 +1,6 @@
-import strawberry
-from aiohttp import hdrs, web
-from strawberry.aiohttp.views import GraphQLView
-from strawberry.types import ExecutionResult, Info
+from strawberry.subscriptions import GRAPHQL_TRANSPORT_WS_PROTOCOL, GRAPHQL_WS_PROTOCOL
 
 from .app import create_app
-
-
-async def test_graphql_query(aiohttp_app_client):
-    query = {
-        "query": """
-            query {
-                hello
-            }
-        """
-    }
-
-    response = await aiohttp_app_client.post("/graphql", json=query)
-    data = await response.json()
-    assert response.status == 200
-    assert data["data"]["hello"] == "strawberry"
 
 
 async def test_graphiql_view(aiohttp_app_client):
@@ -36,111 +18,81 @@ async def test_graphiql_disabled_view(aiohttp_client):
     assert response.status == 404
 
 
-async def test_custom_context(aiohttp_client):
-    class CustomGraphQLView(GraphQLView):
-        async def get_context(self, request: web.Request, response: web.StreamResponse):
-            return {"request": request, "response": response, "custom_value": "Hi!"}
+async def test_turning_off_graphql_ws(aiohttp_client):
+    app = create_app(protocols=[GRAPHQL_TRANSPORT_WS_PROTOCOL])
+    aiohttp_app_client = await aiohttp_client(app)
 
-    @strawberry.type
-    class Query:
-        @strawberry.field
-        def custom_context_value(self, info: Info) -> str:
-            return info.context["custom_value"]
-
-    schema = strawberry.Schema(query=Query)
-
-    app = web.Application()
-    app.router.add_route("*", "/graphql", CustomGraphQLView(schema=schema))
-    client = await aiohttp_client(app)
-
-    query = "{ customContextValue }"
-    resp = await client.post("/graphql", json={"query": query})
-    data = await resp.json()
-
-    assert resp.status == 200
-    assert data["data"] == {"customContextValue": "Hi!"}
+    async with aiohttp_app_client.ws_connect(
+        "/graphql", protocols=[GRAPHQL_WS_PROTOCOL]
+    ) as ws:
+        data = await ws.receive(timeout=2)
+        assert ws.protocol is None
+        assert ws.closed
+        assert ws.close_code == 4406
+        assert data.extra == "Subprotocol not acceptable"
 
 
-async def test_custom_process_result(aiohttp_client):
-    class CustomGraphQLView(GraphQLView):
-        async def process_result(self, request: web.Request, result: ExecutionResult):
-            return {}
+async def test_turning_off_graphql_transport_ws(aiohttp_client):
+    app = create_app(protocols=[GRAPHQL_WS_PROTOCOL])
+    aiohttp_app_client = await aiohttp_client(app)
 
-    @strawberry.type
-    class Query:
-        @strawberry.field
-        def abc(self) -> str:
-            return "ABC"
-
-    schema = strawberry.Schema(query=Query)
-
-    app = web.Application()
-    app.router.add_route("*", "/graphql", CustomGraphQLView(schema=schema))
-    client = await aiohttp_client(app)
-
-    query = "{ abc }"
-    response = await client.post("/graphql", json={"query": query})
-    data = await response.json()
-
-    assert response.status == 200
-    assert data == {}
+    async with aiohttp_app_client.ws_connect(
+        "/graphql", protocols=[GRAPHQL_TRANSPORT_WS_PROTOCOL]
+    ) as ws:
+        data = await ws.receive(timeout=2)
+        assert ws.protocol is None
+        assert ws.closed
+        assert ws.close_code == 4406
+        assert data.extra == "Subprotocol not acceptable"
 
 
-async def test_setting_cookies_via_context(aiohttp_client):
-    @strawberry.type
-    class Query:
-        @strawberry.field
-        def abc(self, info) -> str:
-            info.context["response"].set_cookie("TEST_COOKIE", "TEST_VALUE")
-            return "ABC"
+async def test_turning_off_all_ws_protocols(aiohttp_client):
+    app = create_app(protocols=[])
+    aiohttp_app_client = await aiohttp_client(app)
 
-    schema = strawberry.Schema(query=Query)
+    async with aiohttp_app_client.ws_connect(
+        "/graphql", protocols=[GRAPHQL_TRANSPORT_WS_PROTOCOL]
+    ) as ws:
+        data = await ws.receive(timeout=2)
+        assert ws.protocol is None
+        assert ws.closed
+        assert ws.close_code == 4406
+        assert data.extra == "Subprotocol not acceptable"
 
-    app = web.Application()
-    app.router.add_route("*", "/graphql", GraphQLView(schema=schema))
-    client = await aiohttp_client(app)
-
-    query = "{ abc }"
-    response = await client.post("/graphql", json={"query": query})
-
-    assert response.status == 200
-    assert response.cookies.get("TEST_COOKIE").value == "TEST_VALUE"
-
-
-async def test_malformed_query(aiohttp_app_client):
-    query = {
-        "qwary": """
-            qwary {
-                hello
-            }
-        """
-    }
-
-    response = await aiohttp_app_client.post("/graphql", json=query)
-    reason = await response.text()
-
-    assert response.status == 400
-    assert reason == "400: No GraphQL query found in the request"
+    async with aiohttp_app_client.ws_connect(
+        "/graphql", protocols=[GRAPHQL_WS_PROTOCOL]
+    ) as ws:
+        data = await ws.receive(timeout=2)
+        assert ws.protocol is None
+        assert ws.closed
+        assert ws.close_code == 4406
+        assert data.extra == "Subprotocol not acceptable"
 
 
-async def test_sending_invalid_json_body(aiohttp_app_client):
-    query = "}"
+async def test_unsupported_ws_protocol(aiohttp_client):
+    app = create_app(protocols=[])
+    aiohttp_app_client = await aiohttp_client(app)
 
-    response = await aiohttp_app_client.post(
-        "/graphql", data=query, headers={"content-type": "application/json"}
-    )
-    reason = await response.text()
+    async with aiohttp_app_client.ws_connect(
+        "/graphql", protocols=["imaginary-protocol"]
+    ) as ws:
+        data = await ws.receive(timeout=2)
+        assert ws.protocol is None
+        assert ws.closed
+        assert ws.close_code == 4406
+        assert data.extra == "Subprotocol not acceptable"
 
-    assert response.status == 400
-    assert reason == "400: Unable to parse request body as JSON"
 
+async def test_clients_can_prefer_protocols(aiohttp_client):
+    app = create_app(protocols=[GRAPHQL_WS_PROTOCOL, GRAPHQL_TRANSPORT_WS_PROTOCOL])
+    aiohttp_app_client = await aiohttp_client(app)
 
-async def test_not_allowed_methods(aiohttp_app_client):
-    # The CONNECT method is not allowed, but would require SSL to be tested.
-    not_allowed_methods = hdrs.METH_ALL.difference(
-        {hdrs.METH_GET, hdrs.METH_POST, hdrs.METH_CONNECT}
-    )
+    async with aiohttp_app_client.ws_connect(
+        "/graphql", protocols=[GRAPHQL_TRANSPORT_WS_PROTOCOL, GRAPHQL_WS_PROTOCOL]
+    ) as ws:
+        assert ws.protocol == GRAPHQL_TRANSPORT_WS_PROTOCOL
 
-    for method in not_allowed_methods:
-        response = await aiohttp_app_client.request(method, "/graphql")
-        assert response.status == 405, method
+    async with aiohttp_app_client.ws_connect(
+        "/graphql", protocols=[GRAPHQL_WS_PROTOCOL, GRAPHQL_TRANSPORT_WS_PROTOCOL]
+    ) as ws:
+        assert ws.protocol == GRAPHQL_WS_PROTOCOL
